Refactoring the project structure
