+++ app/src/main/java/com/yurii/youtubemusic/models/VideoItem.kt
@@ -1,0 +1,5 @@
+package com.yurii.youtubemusic.models
+
+import java.io.Serializable
+
+data class VideoItem(val videoId: String) : Serializable
