+++ app/src/main/res/layout/fragment_you_tube_musics.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".fragments.youtubemusic.YouTubeMusicsFragment">
+    tools:context=".fragments.YouTubeMusicsFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
